Added animation for scheduling and different scheduling algorithms.

+++ SmartOS_Core/IOEvent.cpp
@@ -1,6 +1,6 @@
 #include "IOEvent.h"
 
-IOEvent::IOEvent(Type type, unsigned long cycleStamp)
+IOEvent::IOEvent(Type type, size_t cycleStamp)
     : m_type{type}
     , m_cycleStamp{cycleStamp}
 {}
@@ -10,7 +10,7 @@ IOEvent::Type IOEvent::type() const
     return m_type;
 }
 
-unsigned long IOEvent::cycleStamp() const
+size_t IOEvent::cycleStamp() const
 {
     return m_cycleStamp;
 }
